requirements.txt
streamlit
google-genai
firebase-admin
streamlit-mic-recorder